Changes based on code review: Moved update board posting code to be within the files that were calling it, needs further refactor Removed some stray code Fixed some names (assignLogic -> setLogic) Removed some stray comments Refactored a Sequelize call to no longer need a particular import Fixes #264
