+++ source/commands/bounty/addcompleters.js
@@ -2,7 +2,33 @@ const { CommandInteraction, userMention, bold } = require("discord.js");
 const { Sequelize } = require("sequelize");
 const { extractUserIdsFromMentions, listifyEN, commandMention } = require("../../util/textUtil");
 const { addCompleters } = require("../../logic/bounties.js");
-const { updateBoardPosting } = require("../../util/bountyUtil.js");
+
+/**
+ * Updates the board posting for the bounty after adding the completers
+ * @param {Bounty} bounty 
+ * @param {Company} company 
+ * @param {Hunter} poster 
+ * @param {UserId[]} numCompleters 
+ * @param {Guild} guild 
+ */
+async function updateBoardPosting(bounty, company, poster, newCompleterIds, completers, guild) {
+	let postingId = company.bountyBoardId;
+	let { boardId } = bounty;
+	if (!postingId || !boardId) return;
+	let boardsChannel = await guild.channels.fetch(boardId);
+	let post = await boardsChannel.threads.fetch(postingId);
+	if (post.archived) {
+		await thread.setArchived(false, "Unarchived to update posting");
+	}
+	post.edit({ name: bounty.title });
+	let numCompleters = newCompleterIds.length;
+	post.send({ content: `${listifyEN(newCompleterIds.map(id => userMention(id)))} ${numCompleters === 1 ? "has" : "have"} been added as ${numCompleters === 1 ? "a completer" : "completers"} of this bounty! ${congratulationBuilder()}!` });
+	let starterMessage = await post.fetchStarterMessage();
+	starterMessage.edit({
+		embeds: [await bounty.embed(guild, poster.level, company.festivalMultiplierString(), false, company, completers)],
+		components: bounty.generateBountyBoardButtons()
+	});
+}
 
 /**
  * @param {CommandInteraction} interaction
@@ -77,7 +103,7 @@ module.exports = {
 		]
 	},
 	executeSubcommand,
-	assignLogic: (logicBundle) => {
+	setLogic: (logicBundle) => {
 		bounties = logicBundle.bounties;
 	}
 };
