+++ source/util/bountyUtil.js
@@ -46,35 +46,6 @@ async function showcaseBounty(interaction, bountyId, showcaseChannel, isItemShow
 	})
 }
 
-/**
- * 
- * @param {Bounty} bounty 
- * @param {Company} company 
- * @param {Hunter} poster 
- * @param {UserId[]} numCompleters 
- * @param {Guild} guild 
- */
-async function updateBoardPosting(bounty, company, poster, newCompleterIds, completers, guild) {
-	let postingId = company.bountyBoardId;
-	if (!postingId) return;
-	let { boardId } = bounty;
-	if (!boardId) return;
-	let boardsChannel = await guild.channels.fetch(boardId);
-	let post = await boardsChannel.threads.fetch(postingId);
-	if (post.archived) {
-		await thread.setArchived(false, "Unarchived to update posting");
-	}
-	post.edit({ name: bounty.title });
-	let numCompleters = newCompleterIds.length;
-	post.send({ content: `${listifyEN(newCompleterIds.map(id => userMention(id)))} ${numCompleters === 1 ? "has" : "have"} been added as ${numCompleters === 1 ? "a completer" : "completers"} of this bounty! ${congratulationBuilder()}!` });
-	let starterMessage = await post.fetchStarterMessage();
-	starterMessage.edit({
-		embeds: [await bounty.embed(guild, poster.level, company.festivalMultiplierString(), false, company, completers)],
-		components: bounty.generateBountyBoardButtons()
-	});
-}
-
 module.exports = {
-	showcaseBounty,
-	updateBoardPosting
+	showcaseBounty
 }
